Moved config stuff into its own class file
