+++ src/main/java/net/trim02/loginPassword/AdminCommands.java
@@ -1,0 +1,4 @@
+package net.trim02.loginPassword;
+
+public class AdminCommands {
+}
