+++ src/main/resources/config.toml
@@ -35,6 +35,7 @@
 
 [misc]
 	configVersion = "1.4-SNAPSHOT"
+	pluginEnabled = true
 	# set to false to make the login command only accessible to players with the permission node loginpassword.login.
 	# this could be used to have a private server where only some players have access to. oneTimeLogin must be disabled otherwise players with the bypass permissions will always be sent to the private server.
 	loginCommandGrantedToEveryone = true
